Partial conversion to static LocalImage.
We have a cursor which means no calls to resolvers, but only from gallery.  We need to work with just uris from externals.  So it's likely best not to bother assuming we will ever have meta data parsed already.  Just work with uris and make the handful of resolver calls on the fly in the viewer.